Move work-related functions to another path
