+++ src/__test__/components/plots-and-tables/markerHeatmap.test.jsx
@@ -30,7 +30,7 @@ jest.mock('../../../utils/socketConnection', () => ({
     resolve({ emit: jest.fn(), on: jest.fn(), id: '5678' });
   }),
 }));
-jest.mock('../../../utils/cacheRequest', () => ({
+jest.mock('../../../utils/work/fetchWork', () => ({
   fetchWork: jest.fn().mockImplementation((expId, body) => {
     if (body.name === 'ListGenes') {
       return new Promise((resolve) => resolve({
