+++ src/__test__/redux/actions/cellSets/runCellSetsClustering.test.js
@@ -3,12 +3,12 @@ import thunk from 'redux-thunk';
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
 import runCellSetsClustering from '../../../../redux/actions/cellSets/runCellSetsClustering';
 import initialState from '../../../../redux/reducers/cellSets/initialState';
-import { seekFromAPI } from '../../../../utils/seekWorkResponse';
+import { seekFromAPI } from '../../../../utils/work/seekWorkResponse';
 
 enableFetchMocks();
 const mockStore = configureStore([thunk]);
 jest.mock('localforage');
-jest.mock('../../../../utils/cacheRequest');
+jest.mock('../../../../utils/work/fetchWork');
 
 jest.mock('../../../../utils/seekWorkResponse', () => ({
   __esModule: true, // this property makes it work
