+++ src/__test__/redux/actions/genes/loadMarkerGenes.test.js
@@ -3,10 +3,10 @@ import thunk from 'redux-thunk';
 import loadMarkerGenes from '../../../../redux/actions/genes/loadMarkerGenes';
 import { MARKER_GENES_ERROR, MARKER_GENES_LOADED, MARKER_GENES_LOADING } from '../../../../redux/actionTypes/genes';
 import initialState from '../../../../redux/reducers/genes/initialState';
-import { fetchWork } from '../../../../utils/cacheRequest';
+import { fetchWork } from '../../../../utils/work/fetchWork';
 
 jest.mock('localforage');
-jest.mock('../../../../utils/cacheRequest');
+jest.mock('../../../../utils/work/fetchWork');
 
 const mockStore = configureStore([thunk]);
