+++ src/redux/actions/differentialExpression/loadDifferentialExpression.js
@@ -2,7 +2,7 @@ import {
   DIFF_EXPR_LOADING, DIFF_EXPR_LOADED, DIFF_EXPR_ERROR,
 } from '../../actionTypes/differentialExpression';
 
-import { fetchWork } from '../../../utils/cacheRequest';
+import { fetchWork } from '../../../utils/work/fetchWork';
 
 const getCellSetName = (name) => (name?.split('/')[1] || name);
